Convert homework support library to dependence
Originally, the homework support library was small and we wanted to
avoid having to ask students to install dependencies via pip for each
homework. Even when the library grew, it was still copy-pasted into each
assignment repo ("vendored" :) to avoid external dependencies. This made
updates to assignments a pain (and did end up causing regressions on at
least a couple of occasions, unsurprisingly).

However, after switching to DevContainers/Codespaces as the default
setup (and, also, no longer using a conda-based distribution), there is
no reason to suffer anymore. Finally! :)