Configure git pull.rebase in devcontainer
This is a workaround for GitHub Classroom which has what amounts to
a race condition: the Classroom bot takes a while to set up the repo,
so if students clone before it's done, then they encounter issues when
pushing back to GitHub. This is made even more likely by the unfortunate
choice in Classroom to show the "Open in Codespaces" button directly in
the assignment acceptance screen.